Refactor logic to hook
