--- .changeset/perfect-apes-trade.md
@@ -1,6 +1,0 @@
----
-'@toptal/picasso-collapse': patch
-'@toptal/picasso-paper': patch
----
-
-- bump tailwind-merge to 2.3.0
