+++ packages/base/Slider/src/Slider/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './use-label-overlap'
